+++ .travis.yml
@@ -19,6 +19,8 @@ install:
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install numpy; fi
 - if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install fpzip; fi
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install fpzip; fi
+- if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install .dask; fi
+- if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install .dask; fi
 - if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then pip install -e .[boss,test]; fi
 - if [[ $PYTHON_MAJOR_VERSION == 3 ]]; then pip3 install -e .[boss,test]; fi
 - if [[ $PYTHON_MAJOR_VERSION == 2 ]]; then python setup.py sdist; fi
